// src/app/Rest/Conductor/model/bulkResponse.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface BulkResponse {
    bulkErrorResults?: { [key: string]: string; };
    bulkSuccessfulResults?: Array<string>;
}