// src/app/Rest/Conductor/model/workflow.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Task } from './task';
import { WorkflowDef } from './workflowDef';

export interface Workflow { 
    ownerApp?: string;
    createTime?: number;
    updateTime?: number;
    createdBy?: string;
    updatedBy?: string;
    status?: Workflow.StatusEnum;
    endTime?: number;
    workflowId?: string;
    parentWorkflowId?: string;
    parentWorkflowTaskId?: string;
    tasks?: Array<Task>;
    input?: { [key: string]: any; };
    output?: { [key: string]: any; };
    correlationId?: string;
    reRunFromWorkflowId?: string;
    reasonForIncompletion?: string;
    event?: string;
    taskToDomain?: { [key: string]: string; };
    failedReferenceTaskNames?: Array<string>;
    workflowDefinition?: WorkflowDef;
    externalInputPayloadStoragePath?: string;
    externalOutputPayloadStoragePath?: string;
    priority?: number;
    variables?: { [key: string]: any; };
    lastRetriedTime?: number;
    workflowVersion?: number;
    startTime?: number;
    workflowName?: string;
}
export namespace Workflow {
    export type StatusEnum = 'RUNNING' | 'COMPLETED' | 'FAILED' | 'TIMED_OUT' | 'TERMINATED' | 'PAUSED';
    export const StatusEnum = {
        RUNNING: 'RUNNING' as StatusEnum,
        COMPLETED: 'COMPLETED' as StatusEnum,
        FAILED: 'FAILED' as StatusEnum,
        TIMEDOUT: 'TIMED_OUT' as StatusEnum,
        TERMINATED: 'TERMINATED' as StatusEnum,
        PAUSED: 'PAUSED' as StatusEnum
    };
}